did a bunch of stuff
added the thunder image and changed it up to a different file using photoshop to create a web legacy.

Styles on first section need to be added to the other sections once I get it down.

Everything should be 1440 instead of full length desktop screen, so there's white space on the sides.

div class needs to be changed to another name so that div is open