changed the margins and styles within the pages
I've been working on the styles in the pages and working on making them look like how the figma looks. The bottom of the dracula background needs to get cut so Itt can look like the figma design. Doing the sections was harder on my other page as it made everything else go behind the background. Still haven't done the hamburger menu and I fear for it